driver: Split libfuzzer_{driver,fuzz_target} into main and library
By using LLVMFuzzerRunDriver, we can remove the LibfuzzerDriver class
and replace it by:

* a real main function that preprocesses arguments for libFuzzer and
  starts a JVM;
* everything else turned into a proper library with no dependency on our
  custom JVM class.

This will allow us to turn everything except the main function into a
JNI shared library, opening up the possibility to launch Jazzer from an
already running JVM.
