--- driver/libfuzzer_driver.h
@@ -1,59 +1,0 @@
-/*
- * Copyright 2021 Code Intelligence GmbH
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#pragma once
-
-#include <memory>
-#include <string>
-
-#include "jvm_tooling.h"
-
-namespace jazzer {
-
-class AbstractLibfuzzerDriver {
- public:
-  AbstractLibfuzzerDriver(int *argc, char ***argv,
-                          const std::string &usage_string);
-
-  virtual ~AbstractLibfuzzerDriver() = default;
-
-  virtual int TestOneInput(const uint8_t *data, std::size_t size) = 0;
-
-  // A libFuzzer-registered callback that outputs the crashing input, but does
-  // not include a stack trace.
-  static void (*libfuzzer_print_crashing_input_)();
-
- protected:
-  // wrapper around the running jvm instance
-  std::unique_ptr<jazzer::JVM> jvm_;
-
- private:
-  void initJvm(std::string_view executable_path, std::string_view seed);
-};
-
-class LibfuzzerDriver : public AbstractLibfuzzerDriver {
- public:
-  LibfuzzerDriver(int *argc, char ***argv);
-
-  int TestOneInput(const uint8_t *data, std::size_t size) override;
-
-  ~LibfuzzerDriver() override = default;
-
- private:
-  static std::string getUsageString();
-};
-
-}  // namespace jazzer
